skill/openclaw/tavily-search/requirements.txt
tavily-python>=0.5.0
